+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `plugins.mode.web` to control the server-side mode solver.
 - Support for `JaxDataArray.interp`, allowing differentiable linear interpolation.
 - Support for `JaxSimulationData.get_intensity()`, allowing intensity distribution to be differentiated in `adjoint` plugin.
+- Support for `JaxFieldData.flux` to compute differentiable flux value from `FieldData`.
 
 ### Changed
 - Add `Medium2D` to full simulation in tests.